Module Api Def Files Now Support Exporting HttpService Definitions
Now a .claro_module_api file can actually export an HttpService definition as part of a module's API. Consumers of this API can then utilize any of the procedures that Claro autogenerates for the exported HttpService to make calls to the declared downstream service. This is a powerful abstraction that reminds me (in a good way) of grpc's proto service definitions and autogenerated clients.

Additionally, to enable this I fixed some namespacing issues that came up with procedure signatures defined in dep module api files in order to ensure that this change maintains the requirement that to call into procedures defined in dep modules you must prefix the call with the dep module name declared in the relevant claro_module()/claro_target().

To validate that all of this is working I've continued to decompose the demo BuggyBuggies server into more modules that test this functionality.