Massive Overhaul of Existing Progress on Claro's Module System.
Now, in order to avoid the AST nodes and parsers needing to support multiple use cases in handling both direct sources and deps, I've migrated the module system to make use of Protobuf representations of module apis and written logic in the JavaSourceCompilerBackend to manually configure the symbol table with all dep module exported bindings. This change also went too far in that it also adds various other minor features like the ability to export atoms.

Still much more work to be done on the module system but this is a massive reworking of a better foundation to continue making progress atop.