Minor Update to Procedure Proto Representations to Enable Use of Custom Types w/ Generic Type Params in Module Deps
Prior to this there was an internal compiler error if you attempted to initialize a custom type with generic type params if that type came from a dep module. The issue was specifically related to missing information in the proto representation of procedure types where a custom type's constructor type would not be able to correctly reference it's generic type params.