Massive Update to Implement Stdlib Builtin Types as an Implicit claro_module() Module Dep!
This is a major milestone as this is the first time that Claro's builtin Atoms and UserDefinedTypes (so far just `Nothing`, `Error`, and `ParsedJson`) are provided via an actual stdlib Module dep. The dep is included as an implicit dep on all claro_module() targets. Now, rather than directly referencing one of these types like `Error` you reference them like `std::Error`. This is more typing/noise, but in the future that can be mitigated with some syntax sugar for something like `from std using Error;` or some nonsense like that.

This is a big step because it opens the door to me beginning to build out a proper stdlib via specially named Modules. In the future, I think I'll add some requirement that all user-defined dep module names begin with a capital letter so that stdlib modules can be all lowercase and can be added over time w/o worry of conflicting with existing user-defined dep module names.

This is something that can be built upon to start to build out optional stdlib modules for things that are very heavyweight (like an http server) that shouldn't be included by default in every Claro program for codesize reasons (e.g. hello world shouldn't have http server code bundled into it's final Jar).

In a followup CL I'm going to start restructuring the claro_build_rules.bzl extension file so that the claro_module_internal() rule is only accessible to the stdlib module impl directory and not generally made public. A followup will also be necessary to bundle the precompiled stdlib modules with the github releases so that the cli can place deps on it.