Update the claroc.sh Script to Correctly Add Deps on Stdlib Modules
This change packages the registered stdlib modules into the claro_builtin_java_deps_deploy.jar file and then updates the claroc.sh script to dynamically constract --dep and --stdlib_dep args based on the .claro_module files found packaged into the deps jar. This is in an attempt to avoid needing to update this script again repeatedly as I progressively add new stdlib modules over time.

Note to self: this change imposes a restriction on stdlib claro_module() targets that requires that they actually be named exactly as they're used in .claro source files so that the generated <name>.claro_module file is named in a standardized way that can be formatted by the cli in setting the stdlib module dep args.