Support For *Optional* Stdlib Module Deps!
This is the start of a massive win for codesize as Claro is now able to partition it's Stdlib into two portions: one included in all Claro compilation units by default, and the other that's only included on an opt-in basis. This means that "you don't pay for what you don't use".

In particular, this CL makes the first step towards this rich functionality by factoring out Http Service related functionality into the new `http` optional stdlib module and only includes the necessary Java deps in your build if you take a dep on `http` directly. This may seem like I'm just publishing a 3rd party library like any other Claro user will be able to do in the future, but this is different in the sense that users will not know which Bazel target is specifically implementing the module, that is claimed as an internal detail. Users may simply inform the claro_*() rules that they would like to use `http`, and the rules handle assiging the appropriate dependency configuration for you.

Feeling really great about this feature as I've already observed a 30% decrease in binary size for a module that doesn't use `http` now that I was able to avoid placing deps on Http related Java libraries. There's still a lot more dependency refactoring to be done though as the binary size is still way too big for no good reason.