Prototyping IPC Implementation of Dep Module Generic Procedure Monomorphization
This is nowhere near functioning, and so this approach is not yet integrated into the compiler itself. This is simply a prototype that's allowing me to test out the feasibility and robustness of the protocol that would need to be implemented if this were integrated into the compiler. The good thing about this prototype is that the messaging protocol will be fully demonstrated and integrating it into the compiler will mostly involve copying code and working on compilation specific things, and I won't have to worry about the IPC protocol itself if the prototype can be validated.

The prototype itself is not even complete, but it's become sufficiently complex to the point that I fear losing it and so I'm pushing this up now (thankfully I have no users yet so repo health doesn't matter and I can freely push this to main, yolo).

This technique has had one observable consequence on the actual language itself, namely, http client requests no longer have any timeout....this is obviously terrible, so in the future I'll need to update the http module to provide a mechanism for getting a client with a configurable timeout (once the module system is fully supported this will be easy).