Begin Integrating Dep Module Monomorphization Support Into Compiler
This is an extremely hairy CL. I very complicated side-effect of restarting compilation in subprocesses (as well as using Claro's builtin http support) within the compiler's own Dep Module Monomorphization support is that I have somewhat unintentionally entered the world of bootstrapping Claro. So in addition to all of the added complexity of the multiprocessing IPC, I now have to stage changes across releases as local changes won't be available in the "bootstrapping compiler" which is pulled from the latest release on clarolang.com.

For now, I'm pushing this initial commit that triggers recompilation of dep modules, and then ends up erroring out when the subprocess compilation unit attempts to overwrite the .claro_module file of the dep-module. This change includes some staged updates to the JavaSourceCompilerBackend so that after this release the Dep Module Monomorphization subprocess can configure it to avoid writing files. All I want the subprocess to do is to recompile, and then keep AST/Symbol-table in memory for handling monomorphization reqs.