Update the IPC MonomorphizationRequest Proto to Include UserDefinedType Metadata
This is necessary since a generic procedure from a dep module may be called with concrete args that were not present in the dep subgraph of the given module. So, in order for the dep module subprocess to successfully monomorphize for these unknown types, their metadata must be passed along.