Implement Std Lib Module `lists` Providing Operations Over the Builtin Types `[T]` and `mut [T]`
Super exciting step enabling Claro to become that much more useful! This is a porting of the methods provided by the underlying java.util.ArrayList<T> that backs the builtin list types in Claro in the Java codegen. As such, the implementation (and documentation) of all of these procedures comes from https://docs.oracle.com/javase/8/docs/api/java/util/ArrayList.html. In the (distant) future, I'll very likely need some way to automate this process as this wrapping process is very labor-intensive and would be a barrier to making use of existing Java libraries. That future automation should also be done in some way that allows it to be used by Claro users, and not depend on users needing to use inline-Java as this current manual, internal-only approach does.

This module is only possible thanks to the recent (and in-progress) work on dep module generic procedure monomorphization! Starting to pay dividends!