Update `lists` Stdlib Module To Model All Possible Errors as Values
Now, none of the procedures exported by `lists` should be able to Panic at runtime. Instead, all possible errors have been explicitly modeled and callers will be required to either check for them, or explicitly ignore them. This is an effort to establish Claro's philosophical approach to errors as fundamentally values modelled within the type system rather than outside of it as in Java.