Support Exporting Generic Providers From Modules
Interestingly this production is only possible because the stdlib (will in the following CL) use some compiler intrinsics to implement some procedures with this manner of generic provider signature. So, a fun fact is that every procedure w/ this signature validated by the compiler will touch one of those specific stdlib procedures at some point (or else it returns a value that does not actually contain an instance of the generic type (e.g. `provider foo<T>() -> oneof<Other, T> { return Other(...); }`).

Specifically, this is necessary to implement at least the following for now:
  - (futures): provider getImmediateCancelledFuture<T>() -> future<T>;
  - (sets):    provider getEmptySet<T>() -> {T};
  - (sets):    provider getEmptyMutSet<T>() -> mut {T};