Setup Experimental Partial Recompilation for Dep Module Monomorphization
This functionality is being staged in the compiler, guarded behind a check that the compiled module includes "TESTING" in its name so that this doesn't have any effect on existing code until deemed ready/stable. In the following CL, I'll begin utilizing this to only selectively run type checking over generic procedures requested by the monomorphization coordinator. This is going to be a major runtime savings to avoid large amounts of unnecessary work by the dep module monomorphization subprocesses.