Re-add `transitive_dep_module_monomorphization_requests` to `MonomorphizationResponse` Proto w/ Updated Type
This is a migration CL. Unfortunately, working around Claro's multiple compiler versions (latest vs bootstrapping) requires staging changes to even this unused field across multiple CLs. This CL updates the type of the field now that it's free. The next CL will have the dep module subprocess set the field so that the Monomorphization Coordinator can use it in a followup CL.