Fix Maven Deps After Adding protobuf-java-util Dep
protobuf-java-util had a dep on a higher guava version than what I was explicitly depending on and so Bazel's dependency resolution was picking the higher one from protobuf-java-util rather than what I'd explicitly stated. This higher version was the wrong (android) variant that doesn't contain the functionality I needed. So I've updated to the latest Guava version so that Bazel doesn't override it.