First Step Towards Supporting Modules Exporting Contract Defs
This first step just enables the use of the contract def syntax in .claro_module_api files, allows the contract to be used within the module's impl files, and serializes the exported contract in the resulting .claro_module file. This does not yet provide any support for referencing contracts exported by dep modules, exporting procedures that actually require the contract, or for exporting an impl of the contract. Those will come in following CLs.