Serialized Procedures Now Encode the Defining Module Disambiguator for Required Contracts
I'll be migrating away from the old `required_contract_names_to_generic_type_params` fields instead of this in the coming CLs as the new field is utilized. This preliminary work is necessary to separate any naming hackery that the compiler does from the actual .claro_module representation as this should eventually reach a point that it becomes more stable than the compiler implementation itself.