Add a Demo of Emergent Support for "Inheritance" w/o the Need for Subtyping
This is an emergent behavior that I've been pondering for quite some time now and am finally getting around to writing a demo of to validate the functionality! Claro's approach to constructing modules at the Bazel build layer enables Claro to be extremely expressive w/o complicating the language itself, where other languages need lots of keywords and complex semantics in order to provide support. While this is by no means a claim that the "inheritance" possible in Claro is cleaner or easier to use than that available in a classic OO language like Java, I do think that the possibility of this type of expressiveness speaks volumes of the merits of Claro's overall unique design.