Modules Can Now Export Contract Implementations
A .claro_module_api file may now use the following syntax:

```
implement SomeContract<Foo, Bar>;
```

to export a contract implementation and Claro will validate that the declared implementation is actually implemented within the module's srcs.

Included in this CL is the update that a module may implement a locally defined contract over *any* types, whether they were defined in the current module or not. This is safe because every other compilation unit that "knows about" the contract definition itself, will automatically know the contract implementations exported in the same module which avoids any risks of incoherency. This models the other side of Rust's "Orphan Rules" that weren't addressed yet. Again, it's possible that Claro may deviate from Rust's approach in the future, but for now I think it's better to start with a more constrained feature where constraints can potentially be loosened in the future rather than the other way around.