Claro Now Supports Calling Contract Procedures for Impls Defined in Dep Modules
This unfortunately involved a workaround that required the canonical naming of all contracts in-memory to be fully disambiguated at all times so that no matter where the contract is referenced from, the codegen knows unambiguously how to reference the codegen'd procedures. This is going to have a negative impact on error messaging, so I'll have to come back and rethink this approach in the future. At least for now functionality is there.