Include Required Contract Impl Details in MonomorphizationRequest Proto
Some more setup was necessary before the dep module monomorphization subprocesses could make use of the defining module namespacing information for required contract impls. The biggest change here is that the internal Types.$ContractImplementation representation has been updated to include the namespacing information necessary to fully disambiguate where the implementation originated as this information is necessary to codegen the appropriate target.

One subtle bug that got fixed that I'll call out explicitly here, is that the IPC proto serialization encoding between monomorphization coordinator -> subprocess was accidentally using an encoding alphabet that contained invalid chars for use in an http url. This was a simple fix once I discovered it, but now it's important that both sides of the encoding/decoding process use the BaseEncoding.base64Url encoding rather than BaseEncoding.base64 to ensure that this IPC scheme continues to behave properly.

The next CL should hopefully be able to actually start working on the dep module subprocess itself to register required contract impls.