Finish Support for All (non-Dynamic) Contract Procedure Dispatch Involving Dep Modules!
Now Claro should have full support for defining, implementing and using contracts in all sorts of combinations involving crossing dep module boundaries! This means that as long as the current module has a dep on some contract implementation, then that module may call any available procedure requiring that contract!

Next up, I'll need to address dynamic dispatch, as this approach involving definitions in multiple modules currently breaks with the existing (invalid) assumptions that the dynamic dispatch support made in terms of expecting that all implementations would be known within the same compilation unit that the contract was defined in. This is clearly no longer even remotely the case. So, I'll need to backtrack and rework dynamic dispatch in order to fully close out contract support in the presence of modules.