First Pass at AutoGenerated ClaroDocs!
I'm super excited about this prototype coming together in the last couple of days especially as this was my first attempt at doing any sort of web development! So this is an extremely hacky start that I'll largely just tinker on in spare time from here until I can hopefully find someone with real web experience to help me actually turn this into something great.

The intention behind ClaroDocs is to implement some of the usability features that an IDE would normally provide if Claro had LSP support. Since that form of support will very likely be a significant distance off in the future, this static site should provide the ability to jump from type references in module exported procedure args, to the exact definition of the type. It should also ideally provide some mechanisms for traversing the dep graph in various directions. For now, nothing is really supported beyond some basic, buggy rendering of procedure signatures. But this is a great start!