Claro Now Fully Supports Modules Exporting Static Values Initialized on Program Startup!
This is a massive development as it eliminates a large portion of the use cases that dependency injection frameworks get used for! Now, a module can simply declare that it statically exports a static value, and the entire program can use this singleton, deeply-immutable, non-reassignable value anywhere in the program. As a strong argument for the power of this functionality, I've included an update of the Buggy Buggies demo server that takes advantage of this new functionality to allow the use of a *single* statically initialized `HttpClient<BuggyBuggies>` instance throughout the entire server lifetime, whereas prior to this it was impossible for a Claro Server to maintain any state whatsoever, so clients had to be continuously re-initialized on the fly.

To take this example even further, to demonstrate how Claro's static values can eliminate the need for any dependency injection framework, I could use Bazel to select a static value provider at build time based on some flag like `--dev` or `--prod`. See: https://bazel.build/docs/configurable-attributes.