Module's Exported Static Values May Be Declared `Lazy`
For any static value declared to be lazy, as in the following example:

```
lazy static SOME_LAZY_STATIC_VAL: string;
```

the value's associated provider will be called exactly once or never at all. If the value is never referenced at runtime, then it will never be initialized, giving a mechanism for potentially avoiding some slow operation to compute the value in the case where it was never even going to be used (for example reading some large file). The value will be initialized the first time it is referenced at runtime.

Claro does this by codegening a static synchronized method that checks whether the value was already initialized and if not computes and caches the result, and preventing it from being computed again. This synthetic procedure is `synchronized` so that lazy static vals are as safe to access within a concurrent context as eager static vals. This does incur some overhead though, users should be well warned of this in Claro's eventual documentation.

Providers for lazy static values have an additional restriction compared to their non-lazy counterparts - they must be non-blocking. This is because after de-sugaring, referencing a lazy static value is actually a possible procedure call to the lazy static value's provider. If this was referenced within a concurrent context it would break the graph procedure guarantees that deadlocking is impossible. (This could also be addressed by enabling blocking analysis to detect that a reference of this sort in a procedure makes it a blocking procedure, but I don't think I like that idea as switching an exported static value to be a lazy static value would be a breaking change for dependents that would need to mark all their procedures blocking).