Add ExportedFlagDefinitions Field to SerializedClaroModule.proto
This is just a setup CL. This field will be populated with any direct exported flag defs along with the transitive closure of all of this module's deps' exported flags. Claro will validate that flag names are globally unique as they need to be in a single global namespace in order for them to be passed on the cli.