Modules Can Now Export Flag Definitions That Claro Will Auto-Parse From CLI Args!
This is a really exciting step in the direction of configurability and ergonomics as this is a distinct improvement over the way that flags parsing typically works in most (if not all) languages. Traditional languages only give you access to CLI args parsing in the main method/class, and you're unable to configure new flags unless you have access to modify the main method and then manually save it's value somewhere centrally to use it where you actually need it. Now, in Claro (in yet another example of direct inspiration from Google's internal Boq framework) arbitrary modules can declare a new flag that Claro will parse automatically without any changes being needed anywhere except for in the module that wants to make use of the flag.

Importantly, at the moment, I don't plan to allow flags to be hidden internally within modules. If you're writing some library that parses flags to modify behavior then that should be made explicit to users.

Implementation Limitation: Note that in order for this feature to interact properly with the recently implemented static value exports feature, the flags had to be fetched via a JVM System Property 'sun.java.command'....this is actually a potential portability limiter in that it's not guaranteed that every JVM implementation actually supports this System Property... For now, it's working in the JVM I'm using, and it appears that there's at least fairly broad support for this across major JVMs, so I'm going to punt complicating this down the road if portability concerns ever come up from any real user(s).