Add Initial Support for Modules Exporting Opaque Type Defs
This is going to be a powerful encapsulation feature that allows Modules to hide the definition of a custom type's wrapped value type so that it's only known within the implementation of the module. This will serve to define types that may only be instantiated *AND* used by the module that defined it. Other consumers of the module may receive instances from the defining module, but they will only be able to pass them around or aggregate them in some way, and in order to "do" anything with the values, they'll need to be handed back to the defining module.

For now, this CL is just setup. You can now define opaque types in a .claro_module_api file, and Claro will validate that they're correctly defined within the module's impl srcs. However, they can only be used by the defining module. A followup CL will ensure that the defining module actually includes these opaque types in the serialized .claro_module and then imports them during dep module setup.