Claro Supports Statically Validated Resource Files!
In my eyes this is a huge achievement building off of the latest support for opaque types. Now Claro supports a synthetic module `resources` which can be used to access a resource file listed in the build target as
```
resources = { "MyFile" : "some_file.txt", }
```

simply by writing the static value reference `resources::MyFile`. This should ensure that reading static files is trivially straightforward in Claro, whereas in my experience this has been a major source of pain for me in Java. Just figuring out where Java's going to put these files in the JAR is very confusing. Thankfully, within the Bazel build world, there's a consistent answer to where these files will be located, so I can codegen the appropriate resource path.

Still need to followup and actually include the new `files` module in the stdlib.