Begin Attempted Migration to Bzlmod
This first step is just a duplication of the WORKSPACE file to the new WORKSPACE.bzlmod and the addition of the new MODULE.bazel file. This approach means I'm essentially not actually using any of the improved declarative bazel_dep() support that MODULE.bazel offers, but I'll try to do that in a follow-up. For now, this trivial first step is just to get the marker files in place to see if it's possible to test registering Claro in the BCR in a personal fork.

If registration in the BCR can actually work, then users could try out Claro with only a Bazel install and a bazel_dep(name = "claro-lang") in their MODULE.bazel.