Update MonomorphizationCoordinator.java to Bundle dep_module_monomorphization_deploy.jar in JAR
This fallback will be used instead of the Runfiles version when it's not available. This will be the case for the sake of the non-dev usage of the prebuilt compiler outside of this Bazel project. This simply enables non-Claro-compiler-devs to build Claro projects without needing to rebuild the compiler from source every time they do a clean build.