Make BUILD Targets Publically Visible to Enable Downstream Projects to Build From Source
This is intended to be a temporary change as I don't love the idea that downstream projects can take critical dependencies on Claro's internal details so easily. But for now, until I'm able to find some way to get downstream projects to be able to use a prebuilt compiler binary, this is really my only option.