Ready Claro Project for Adding to BCR
This is a large CL but it mostly moves the example claro programs from //src/java/com/claro/claro_programs/... to //examples/claro_programs and updates build references as appropriate.

I've then added a test Bazel module at //examples/bzlmod/... that can be later included in the presubmit.yml that I'll upload with claro-lang's BCR pull request.

Lastly I've made significant updates to the README so that potential users will be able to learn the current steps required to work with Claro and create their own project.

Lastly, update the Bzlmod version to 0.1.316. This should ideally be the first version pushed to the BCR if everything goes to plan.