Bump Bzlmod Version in Example Bzlmod Project
This unfortunately requires bumping the Bzlmod version of the entire Claro project so that I can avoid patching the MODULE.bazel file submitted to the BCR. I really need to develop some standardized process for creating a BCR-ready release. At least a test that validates that both MODULE.bazel and examples/bzlmod/MODULE.bazel agree on the version to be released.