Add Support for Thread Safe Cache in Optional Stdlib
This is a wrapping of the caffeine cache at https://github.com/ben-manes/caffeine. This CL exposes the core caching functionality, but doesn't yet expose a custom CacheLoader that can be used to customize the bulk load and reload operations.

This goes to demonstrate that Claro's stdlib can be used to provide threadsafe mutability that can be used from a graph procedure context in a way that the Claro language mutability validations will approve of. The introduction of this module essentially changes the semantic definition of the `mut` keyword from "objects of this type are mutable" to "objects of this type are mutable but those mutations aren't guaranteed to be threadsafe".