Demonstrate "Parameterized Modules" / Implement custom_cache_loader.bzl's cache_with_custom_loader() Macro As Part of the "Stdlib"
This is a demonstration of Claro's powerful support for extremely flexible build-time metaprogramming as a result of it being fundamentally intertwined with the Bazel build system. With this build-time metaprogramming, Claro programs can relatively easily create so-called "Parameterized Modules" (if I were to use OCaml's terminology for this type of thing - although I only learned of OCaml's support for that after designing Claro's module system).

The fascinating thing about this approach is that this literally generates a wholly new module for each set of params passed to the cache_with_custom_loader() Bazel macro. Pre-existing Bazel functionality is being repurposed here to extend Claro very far beyond where the core language goes within .claro files. And I argue that this is a GOOD thing following my overall design goals for the language! Languages should be *aware of the fact they're being used within a Build System* and take advantage of that fact in an intrinsic way!