Update Types.UserDefinedType toString() Representation to Include Defining Module
While this is a convenience to make it easy to use `type(myUserDefinedTypeInstance);` to print a debug string with a clickable direct link to the defining module BUILD target, it is actually far more significant than that.

This toString() representation of a type is used to generate the names of generic procedure monomorphizations (to encode the concrete types of the monomorphization). So, what this actually does is ensure that when there are two monomorphizations over distinct types with the same name but originating from different defining modules, the monomorphic procedure names will actually be unique. Before this the generated names would conflict because only the UserDefinedType's name was being used in the codegen. This had the effect of actually making it possible for codegen to incorrectly produce calls to the WRONG implementation not intended by the user's code. So this is a significant fix.

NOTE: This is unfortunately going to be a breaking change with the effect that after this push, it's guaranteed that all Claro programs that make use of dep module monomorphization will be broken as the bootstrapping compiler won't yet be configured to use this updated naming convention. So it'll require a second push to simply update the bootstrapping compiler to point to the release containing THIS CL. Hopefully it'll be possible to produce atomic releases that avoid this sort of breaking change at some point in the future.