Support "Abstract Modules"!
Yet another demonstration of the powerful integration between Claro and Bazel.

"Abstract Modules" offer a powerful expressiveness to the Claro ecosystem by leveraging Bazel to handle complex code organization and reuse responsibilities at Build time PRIOR to Compile time. This demonstration effectively adds support for the organizational power of OO-style "Classes" to Claro entirely using Build level tooling and not touching a single line of internal Claro compiler code.

An Abstract Module allows an experienced Claro developer to produce effectively a Module-constructor (referred to as "functors" in some ML-family languages like OCaml/F#/Futhark) by declaring a module api, type parameters, and some optional default or overridable srcs, and deps. Consumers of this Module-constructor can then declare their own concrete Modules by simply providing overrides and/or their own optional extra srcs. I say this will be done by an "experienced Claro developer" b/c this is admittedly not as straightforward as a syntax-supported "Class" system in a Java-like language as it's DIY approach does require some more fundamental level of understanding of what a "Class" *really* is in order to be able to make use of this effectively.

The high level argument that I'm making with this side of Claro is something along the lines of claiming that languages should truly do more to acknowledge that build systems are a fundamental aspect of interacting with a programming language, and they can be leveraged for great gain if the language doesn't fight against it.