Improve Procedure Calls to Dispatch to Static Procedures in Java Codegen
...when not calling through a first-class procedure reference.

This should be a relatively significant performance improvement as now in the general case the vast majority of procedure calls will be direct calls to static methods. Previously all calls were multiple steps of indirection away from the actual logic as in the java codegen they were deferring to a first class ClaroFunction wrapping the actual procedure in every case, not just when calling through a first-class procedure reference. Now, that cost is only paid when you've opted in by passing the procedure as a reference.

NOTE: This is another breaking change for any program currently using dep module monomorphization as this changes Claro's calling convention. This should be fixed after the bootstrapping compiler can be updated to include this CL.