Prime the Bootstrapping Compiler for Stdlib Relocation
Automatic Error Propagation (via `?=` operator) requires that the rhs be a `oneof<..., std::Error<...>>`, but after the move it's unable to validate the expected defining module disambiguator so the current stdlib usages of the operator are being rejected for being the "wrong" Error type.

Updating the StdLibModuleRegistry.java to reference the correct new location for the relocated stdlib so that the rebuilt bootstrapping compiler with this change will be able to build the stdlib.