Update Example Bzlmod Project's .bazelrc to Run Via Remote JDK 11 by Default
Of course, users would be free to override this to choose any JDK version they'd prefer, but documenting this default for users gives them a clear path to the single install necessary to be up and running with Claro being Bazelisk itself. By allowing Bazel to download and use this default, then potential Claro users wouldn't have to worry about the headache of getting the correct Java version installed locally before getting started with Claro development.